Separated folders for CSS and JS
