+++ src/components/CSS/ManualReg.css
