+++ src/components/CSS/ManualRegStudentDetails.css
