+++ src/components/JS/Dashboard.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import {  useNavigate } from 'react-router-dom';
-import './Dashboard.css';
+import '../CSS/Dashboard.css';
 
 function Dashboard() {
   const navigate = useNavigate();
