+++ src/components/JS/SchoolLoginPage.js
@@ -1,16 +1,16 @@
 import React from 'react';
 import { useNavigate } from "react-router-dom";
-import './SchoolLoginPage.css'; 
+import '../CSS/SchoolLoginPage.css'; 
 
 
-const AdminLogin = () => {
+const SchoolLoginPage = () => {
   const navigate = useNavigate()
   return (
     <div className="login-container">
       <div className="login-form">
         <h1>Welcome!</h1>  
         <h3>Sign in to</h3>
-        <p> the Admin Login portal</p>
+        <p> the School Login portal</p>
         <div>
         <form className='container'> 
           <div className="form-group">
@@ -29,7 +29,7 @@ const AdminLogin = () => {
           <button type="submit" onClick={(e) => {
               e.preventDefault(); 
               const schoolName = "Apeejay School, NOIDA"; // sample school name
-              navigate('/admin-form', { state: { schoolName } });}}>Login</button>
+              navigate('/dashboard', { state: { schoolName } });}}>Login</button>
         </form>
         </div>
       </div>
@@ -40,4 +40,4 @@ const AdminLogin = () => {
   );
 };
 
-export default AdminLogin;
+export default SchoolLoginPage;
